[nfc] Move convert code all inside its entry-point function
This makes it easy to start keeping some state around throughout the
handling of a given source file (which will also be a helpful building
block in doing so throughout the handling of a whole program.)

This nested-functions style is all over the TS implementation.
It seems helpful there; try it out here.