deps: Upgrade recast to my branch fixing Flow function types
I've sent the fix as a PR:
  https://github.com/benjamn/recast/pull/1089
which hopefully will get merged in due course.  In the meantime,
start using it.

The library needs a build step.  Oddly it gives a couple of TS
warnings, which it doesn't when I build the same version in its own
worktree.  Perhaps `tsc` is letting this project's tsconfig leak in,
because it's at an ancestor directory?  Anyway, just ignore for now,
with `|| :`.