vscode: Set tab size to 2, matching Prettier
The default for this was 4.

This change causes copy-pasted code, in particular, to get indented
correctly from the beginning -- rather than having its indentation
messed up until running the auto-formatter.

(This is configurable on the Prettier side too, so we could always
make them agree on 4 instead.  But I've been happy with 2.)