convert [nfc]: Factor out convertMembers
We'll want to share much the same logic between handling the members
of a class-like declaration (i.e. a class or interface) and those of
a type literal (the thing Flow calls an object type.)

At present we have much more complete logic for the former, so start
by just pulling that logic out into a function we can share.