mapper: React.Component, partly
And test it in both a type reference and a class-extends context.

This is our first rewrite of something imported (vs. something that
comes from the TS default library.)

For the moment, we handle the `import { Component } from 'react'`
case, where it's imported individually, but not the case of
`import React from 'react'; … React.Component`, where the module
is imported as a whole.  We'll get to that next.