mapper: Look in interface-extends, too, for needed rewrites
It'd currently be annoying to write a test demonstrating the bug this
fixes, because we currently only rewrite default-library symbols, and
those are global across the whole program and therefore the whole
unit-test suite.  As a result, as long as a given symbol appears
somewhere as a type reference (which it does for the sake of testing
it in that context), it won't trigger this bug anywhere.

But soon we'll start rewriting imported symbols, in particular from
React.  Those symbols will be per-file, and indeed per-import.  So
that will make it easy to demonstrate.