mapper: React.Component, the other half
This covers the `import React from 'react'; … React.Component` case
(and `import * as React …`), to go along with the case of
`import { Component } from 'react'`.

Happily, this involves no changes that know anything at all about the
particular module or name we're rewriting -- it's all in general
infrastructure.  So for the next thing we add a rewrite for, it'll
apply to each of these forms automatically.