Rely on ES2021, requiring Node 16+
Node 16 is the active LTS:
  https://nodejs.org/en/about/releases/
So even for a mature project it'd be reasonable to require it; for a
new project like TsFlower, it's downright conservative.

Requiring that lets us rely on ES2021 (according to this resource:
  https://github.com/tsconfig/bases/blob/main/bases/node16.json
linked from the TS docs at:
  https://www.typescriptlang.org/tsconfig#target
), which gets us useful features like `String#replaceAll`.

Node 18 is out, and will be the active LTS in a few months, so it'd
also be reasonable to require that.  That would get us ES2022.