mapper: Canonicalize filenames to absolute for comparison
The values of `sourceFile.fileName` can be a mix of paths that are
verbatim from what we passed to `ts.createProgram` -- so potentially
relative -- and paths that are absolute, even for a file that we did
pass to `ts.createProgram` and gave a relative path for.  (The latter
seem to come when TS finds that file first via an import.)  That can
cause these comparisons to fail, so that we miss processing a file.

To fix that, canonicalize all the filenames by making them absolute.