convert [nfc]: Fix "as" lines on switch in convertMembers
These grew additional `case` lines a while ago, but the casts didn't
get expanded to match.  The use of `switch (node.kind)` like this, and
`as` at each case, is an idiom we took from the TS implementation --
but it really is a pretty regrettable one for a type-checker, of all
things, to have.  :-/