convert: Translate TS type literals as inexact object types, not exact
This mirrors the actual semantics of the TS object types we're
translating.

It's also essential in order for the very common TS idiom of
intersections on object types to work as intended.

There may well be some cases where an exact object type would be
more helpful -- like the props of a React element, which was the
original motivating use case for Flow to introduce exact object
types in the first place.  We may try to introduce some heuristic
in the future to identify those.  But it seems like the right
baseline should be to faithfully translate the input's semantics.