rewrite: JSX.Element, found as a global augmentation
This is defined in `@types/react`.

One hopes nobody is working with a different global `JSX.Element`
augmentation that's incompatible with this one!  It'd be a mess if
they did, long before anyone tried to run TsFlower on that code --
at least it would if it's ever used together with React.