rewrite [nfc]: Move type definitions here from mapper
These were in mapper.ts because `MapResult` is the type that the
mapper returns when the converter queries it.

But much of what's going on with these types is really about the data
structure of our rewrites, and this is the file where those live.
Ultimately the mapper might end up returning something slightly
different from what it consumes, in which case we'll describe those
there, but for now just move the types here.

In particular, this way the mapper depends on the rewrites (as it
must), but no longer vice versa.