lint: Actually successfully enable Google style
The `eslint-config-google` package is intended to be applied after
either `eslint:recommended`, or some other similar base config.  In
particular, it omits lots of rules for things that are obviously bad.
(See the comment at the top of it.)

So do that.  Seems like `yarn create @eslint/config` went astray here.

As before, leave any actual code changes for separate commits.