convert [nfc]: Inline convertTypeReferenceLike at both call sites
As we add more kinds of rewriter, we'll want the logic at these two
sites to diverge.  In particular at convertTypeHeritage the syntax
only allows a dotted-name and type arguments (and in fact for
class-extends, in the future when we correctly treat it as a hybrid
type/value reference, the rewrite will only provide type arguments),
while at convertTypeReference we can rewrite to an arbitrary type.