convert [nfc]: Separate cases for interface-extends and class-extends
We've been treating these the same, but they should actually be
quite different.  Duplicate the code, now that we've shortened it by
factoring most of it out, so that it'll be straightforward to change
the one that's wrong.  (Though we may not do that immediately; the
cases where the logic we have gets a wrong answer are uncommon.)

Conversely, we've had class-implements unimplemented, but it should
actually get just the same logic we have for interface-extends.