test: Attempt to demo class-extends hybridity
I think to get an interesting test here external translation may be
required -- otherwise the multiple declarations of the base produce
a Flow error themselves.

Then I tried here to demo the issue by exercising the logic about
needing an empty list of type arguments instead of an absent one.
But it seems that when it comes to class-extends, Flow stops caring
about that rule!  So that won't work.

(Maybe it'd start enforcing it after all if I pushed the
consequences more?  Like actually make it matter what the type
parameter gets set to.)

That suggests that the only way to exercise this issue may be on
something we have a rewrite rule for.  At that point, I think I'm
content to let the issue sit there for now, and revisit upon
actually seeing it come up in practice.