WIP some puzzling debugging
(In the original version of this commit, the import at the top was a
default-import `import React as …` rather than a namespace-import
`import * as React …`.)

Ultimately it seems like `React.Component` isn't getting resolved;
like the import isn't getting resolved.  So e.g. `getTypeAtLocation`
and `getSymbolAtLocation` return error or undefined on it.