deps: Add react-native as dev-dep... and oof, its dependencies
We'll use this for Flow to type-check the output of tests where we're
rewriting imports from react-native.

But oof, this adds a lot of transitive dependencies.  Really we'd be
just fine with react-native itself and *not* its dependencies, other
than React and possibly a few others; overwhelmingly they're not
things that are used *in* a React Native app, rather they're part of
the build-and-dev system for it, and consequently they're not things
that get transitively imported by app code or appear in the types
that React Native itself declares.

There's probably a way to arrange that in Yarn.  If not, we could
always make a variant react-native build, with the dependencies
knocked out, and publish that on NPM; as a bonus, it could drop all
the native code too.  For now, just live with it.