convert [nfc]: Standardize name of convertTypeHeritage
This takes a fragment of the input AST, and produces some
corresponding output AST.  We've been calling such functions
`convertFoo`.

When I wrote this I shied away from using that naming, because its
return value isn't literally an AST node.  But it just contains some
AST nodes, with a bit of structure around them.  We have a number of
other functions like that, and have used the same naming convention
for them.  Do so here too.