subst: Move React.FunctionComponent; fix test's import to namespace-import
This does introduce a bit of a regression: if (a) a reference to
this type goes through a default-import (`import React …`) rather
than a namespace-import (`import * as React …`), and (b) it doesn't
pass a type argument (relying on the default), then the default
import defeats our logic in `convertTypeArguments` to detect that
type arguments are needed and give it an empty list `<>` of them.
We end up with a bare reference with no type arguments, which is a
Flow error when the type is generic.

Previously, this was papered over by our handling of
React.FunctionComponent being a full-blown macro.

We should fix that eventually, but for now just note the issue and
move on.  Default imports of React seem a lot less widespread than
namespace imports; perhaps we're not the only tool in the TS
ecosystem that has trouble with them.  And this was quite a lot of
code for what's fundamentally a two-line Flow definition.