deps: Drop flow-parser
We were using this to parse fragments of Flow code that described
substitutions.  We're no longer doing that -- all substitutions are
import substitutions (from `prepImportSubstitute`), where what we emit
has a fixed form for which it's easy to just build the AST directly.
So drop the dependency.

We might well want this again in the future, e.g. in order to parse
those substitution files.  Easy to re-introduce the dependency then.