integration: Leave out react-native-gesture-handler for now
The main reason this was included is that it's a dependency of
react-navigation, so to completely handle the latter's types we'll
need to handle the former.

But there's only a handful of places where react-navigation actually
refers to it, and not deep in its core types.  So a good
intermediate goal would be to allow those imports to be untyped,
while handling the rest of react-navigation.  And in fact at this
point react-native-gesture-handler, despite being smaller, is the
large majority of our remaining Flow errors in the integration suite.