subst/react-native: View, Text, and other component instance types
The types found in `@types` for these look a lot *like* what's
actually found in RN, but are subtly different.  In particular,
in RN they're not types, so that trying to use them as types
produces errors like:

  Cannot use `View` as a type. A name can be used as a type only if
  it refers to a type, interface, class, or enum definition. To get
  the type of a non-class value, use `typeof`. [value-as-type]

Fix that by rewriting such references to refer to the intended types.