convert: Keep value imports even when substituting type
No unit test in this commit, because we don't currently have a good
setup for testing the behavior of SubstituteType short of having
something in our actual substitutions that exercises the behavior.

We'll shortly add some react-native substitutions that depend on
this change, so the tests for that will effectively test this too.

In the future when we have a way for a TsFlower-using project to
specify its own particular substitutions, that will provide a
natural way to unit-test the details of SubstituteType in general.