convert [nfc]: Push `import type` modifiers down to individual items
This will make it easier to convert some of them to `typeof`.

If we were acting on runtime code instead of type definitions, I
think this might alter the semantics.  Just leave a todo-runtime
comment about that; if we someday do try to handle runtime code,
we can deal with it then.