convert: `import type` of a value, emitting `import typeof`
This resolves 8 errors in the integration suite.

Along the way, it resolves 2 more: we drop imports that are of only
a namespace, if they're in `import type`.  We should still drop them
when they weren't marked as type-only, but that doesn't occur in the
integration suite, so demote it out of that list.

On the other hand, 6 new errors appear, of two different kinds.