docs: Better diagnose a pair of new missing-type-arg errors
The issue isn't really that we ought to be emitting an empty
type-argument list here; we shouldn't have to.

Moreover, if we did, it wouldn't work around the issue: these
particular generic functions don't have defaults for their type
parameters, so there'd still be an error unless we specified
particular types for the arguments.  (We could try the bound,
I guess; but it's not clear the resulting type would actually
work anyway.)