mapper [nfc]: Factor out propagateTypeRenames
This will be helpful in letting us similarly propagate namespace
rewrites -- it'll let us keep using early-return control flow for the
parts that are specific to the type rewrite, without affecting any
namespace rewrite.

The one algorithmic change here is that we're a bit later in bailing
out when we already have a type-rewrite for the local symbol: we now
do that only after we've already gone and looked up the imported
symbol.  That wasn't a significant optimization, because we're here
walking *all* imports, and typically most of those will never
participate in a rewrite.